fix ptr

+++ jacobi.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void jacobistep(float *psinew, double *psi, int m, int n);
+void jacobistep(float *psinew, float *psi, int m, int n);
 
 double deltasq(double *newarr, double *oldarr, int m, int n);
